+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 #### 2.5.0 (To be released)
 
+* Add config property `hawtio.localAddressProbing`. Now you can disable local address probing
+  for proxy whitelist upon startup by setting this property to `false` (default: `true`).
+
 #### 2.4.0
 
 * Support Java 11